Replace numpath tests with geomstr
